Replace main and cli entrypoints with files with babel-register hooks
