+++ lib/cli-index.js
@@ -1,0 +1,3 @@
+require('babel-register');
+
+require('./cli');
