+++ lib/cli.js
@@ -7,7 +7,7 @@
 
 var pkg = require('../package.json');
 var cli = require('commander');
-var Importer = require('./index');
+var Importer = require('./importer');
 
 cli
     .version(pkg.version)
