+++ package.json
@@ -46,7 +46,7 @@
   },
   "license": "MIT",
   "bin": {
-    "gremlin-importer": "lib/cli.js"
+    "gremlin-importer": "lib/cli-index.js"
   },
   "dependencies": {
     "babel-preset-es2015": "^6.6.0",
